feat: switch GraphQL fields to Long and introduce DateTime scalar for Instant (#2493)
Agent-Logs-Url: https://github.com/orange-buffalo/simple-accounting/sessions/82bf5980-b71e-4fe7-a06d-6cbafb6756c1

+++ app/src/main/kotlin/io/orangebuffalo/simpleaccounting/business/api/IncomeTaxPaymentsGqlApi.kt
@@ -8,7 +8,7 @@ import java.time.LocalDate
 @GraphQLDescription("An income tax payment in a workspace.")
 data class IncomeTaxPaymentGqlDto(
     @GraphQLDescription("ID of the income tax payment.")
-    val id: Int,
+    val id: Long,
 
     @GraphQLDescription("Title of the income tax payment.")
     val title: String,
@@ -26,5 +26,5 @@ data class IncomeTaxPaymentGqlDto(
     val notes: String?,
 
     @GraphQLDescription("IDs of documents attached to this income tax payment.")
-    val attachments: List<Int>,
+    val attachments: List<Long>,
 )
